removed the dtl directory
